adding delete route, refactor products code modify products db
